Modify the contents of the whole site
These modifications comes as an inspiration of the sites from fellow students who did this assignment on the odin project (TOP). Moved by how good there sites were, I had to improve my site but still keeping in mind it should have basics stuffs of HTML and CSS. I wanted to commit on every change that I make but I saw it hectic since it a personal practice site not a real project or a collaboration work.